\name{mf}
\alias{mf}
\title{Make package function}
\usage{
mf("name")
}
\description{
This function calls a python script to make the function file and the complementary documentation file in the R package. It opens the function to edit afterwards. This is a development tool.
}
\examples{
mf("new_function")\n mf()
}
